requirements.txt
transformers>=5.0.0
torchaudio>=2.5
safetensors>=0.4
tiktoken>=0.7
einops>=0.8
psutil>=5.9
numpy>=1.26
soundfile>=0.12.1
PyYAML>=6.0
hf_transfer
gradio>=5.0.0
torchcodec
